Remove setup function
